// client/src/api/model/limitType.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */

export type LimitType = typeof LimitType[keyof typeof LimitType];


export const LimitType = {
  None: 'None',
  Per_Round: 'Per Round',
  Per_Encounter: 'Per Encounter',
  Per_Session: 'Per Session',
} as const;